move allocation profiling into own mod
