+++ profiling/build.rs
@@ -361,6 +361,7 @@ fn cfg_php_feature_flags(vernum: u64) {
     if vernum >= 80400 {
         println!("cargo:rustc-cfg=php_frameless");
         println!("cargo:rustc-cfg=php_opcache_restart_hook");
+        println!("cargo:rustc-cfg=php_new_zendmm_hooks");
     }
 }
